choose the most recent archive (fixes #3)
If the archive being extracted from is generated from tar_scm with the
versionformat parameter including the git SHA (%h) or timestamp (%ct)
then most time it runs it will get a different name. This means that
extract_file requires wildcard parameters:

  <service name="extract_file" mode="disabled">
    <param name="archive">containment-rpm-0.23+git.*.tar.bz2</param>
    <param name="file">*/containment-rpm.spec</param>
  </service>

This works fine the first time, but on subsequent calls, the working
directory contains multiple archive files. At this point:

  existing_archive="$PWD/$(echo $MYARCHIVE)"

results in $existing_archive containing all of the archive filenames,
not just the newest, and then the archive extraction fails.

Instead, make it choose just the newest.